Implement the Server Bootstrap from CLI
As an optional functionality, the CLI can bootstrap a remote (or
localhost) Kaprien Server from the Ceremony.

It will, before starting the Ceremony, if the server chosen is availble
for bootstrap, than proceed with Ceremony steps, and in the end it will
ask last time if wants to proceed with the Server bootstrap and post to
the bootstrap endpoint the Payload.

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>