Project and Component renaming
This project is moving from Kaprien to a new name, TUF Repository
Service (TRS).

This component moves from `kaprien-rest-cli` to
`tuf-repository-service-cli`.

License move from Apache-2 to MIT.

It started as a personal summer project and will now be an Open
Source Project. 🚀

Thanks to my family (Enzo and Priscila) for all your support. ❤️

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>