Fix the documentation for rstuf command
After the `trs-cli` renaming, it becomes `rstuf`.

For the previous project name TUF Repository Service (TRS) the name
`trs` was named `trs-cli` to avoid conflict with Linux `trs` command
for filer replacing stings.

The project's Repository Service for TUF (RSTUF) simplifies to
`rstuf`. The code and tests are correct, but the documentation was
wrong.

Closes: #77

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>