Run isort with the pre-commit tool
In the .pre-commit-config.yaml file when we call isort we pass the
current directory as a location where isort should lint.
This in turn overrides all excludes made by the pre-commit tool as after
that, we call isort upon all folders.

For example, I have a local virtual environment with the name ".env" and
even if I add the "exclude" option for isort it doesn't work.
Now, when we don't provide the target folders to isort the pre-commit
tool does that and successfully ignores the ".env" folder.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>