Add `crypto` extra to the `securesystemslib` requirement
The `cryptography` package is a dependency of the `securesystemslib`
package, to be exact it is an _optional_ dependency.
`securesystemslib` uses "optional dependencies" ("extras") to allow
a default installation with limited capabilities.
However, since we are making use of the `cryptography` library
via `securesystemslib`, we need to install at least the `crypto`
extra.

Fixes #131

Signed-off-by: Konstantinos Papadopoulos <konpap1996@yahoo.com>