Add `mypy` to `pre-commit` and to `gh-actions`
Since the official mypy repo doesn't provide a pre-commit hook
we have to resort to mirrors-mypy which runs mypy from an isolated
virtualenv and is not as flexible as mypy (e.g., cannot exclude
specific modules as we do in the pyproject.toml mypy config).
For this reason we are creating a local hook to run mypy from.