Fix ModuleNotFoundError: No module 'sqlalchemy'
We have made "sqlalchemy" an optional dependency only for users who
want to use the "import-targets" CLI command.
The reason for that is that we expect that this command will be rarely
used.
For that reason, "sqlalchemy" is not installed by default, but
we import all modules within the "repository_service_tuf/cli/admin"
folder and given we imported "sqlalchemy" globally in import_targets
module this resulted in an error "ModuleNotFoundError:
No module 'sqlalchemy".

To resolve this issue we decided to import "sqlalchemy" inside the
import_targets() function so that only when the "import-targets"
command is used then you would see you will possibly get this error.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>