Ceremony: remove metadata creation except for root
Remove the creation of all metadata files in the bootstrap ceremony
except for root.
We have decided that given that all roles except root share the same
online key, then there is no need to generate metadata for timestamp,
snapshot, targets and all delegated bins.
Instead, we can generate metadata only for root and inside RSTUF
worker we would create the rest of the metadata files as we would have
access to the online key.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>