Add parameter `--no-auth` to the CLI
This adds the `rstuf --no-auth` parameter.

This parameter skips authentication checks/routines when communicating
with RSTUF API. This makes RSTUF CLI compatible when the user decides to
disable RSTUF API built-in authentication.

Added a small refactoring about using the rich console and in shared
calls in the helper.
Added refactoring for the CLI init test file.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>