Trigger FT when PR is approved
This commits a new workflow.
When a PR is approved, it will trigger the FT tests.

In this way, before merging, the PR maintainers can check how the PR
changes the functionality in all RSTUF components.

The failure of this test is not a non-go for merge, but it shows
possible problems.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>