Bootstrap: add key names and do a small refactor
Add a question asking the user if he wants to provide a name for his
key.
This will be useful when showing information about roles later during
the validation phase.
It will be also useful when showing the keys information during metadata
update.

Additionally, I noticed that if I make BootstrapSetup.root_keys as a
dict instead of a list this could simplify some of our existing
functions and a new one that I added when asking for key names.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>