Change CLI to not use built-in auth by default
As implemented the task which makes the RSTUF API
Authentication/Authorization disabled by default
(https://github.com/repository-service-tuf/repository-service-tuf-api/issues/282)
the default behavior of RSTUF CLI to not use auth by default.

If the user wants to use the CLI to the API with built-in
authentication, it requires the parameter `--auth`.

`rstuf --auth admin`

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>