Improve signing keys handling
There is a problem in the implementation until now that while we check
if we have enough singing keys when the users want to stop adding keys
in "_keys_additions" then we check the number of keys against
threshold without taking into account that not all of those keys are
loaded. This means they cannot be used for signing and the check is
wrong.

I also improved generate_payload in order to make it clearer with which
keys we do sign at which stage.

Finally, I improved the documentation for the user, so that it's clear
that when a user considers if he wants to add a key he must take into account
the loaded keys that he can use for signing and not all of the root
keys.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>