If the threshold is not met, don't ask the user
If the threshold requirement is not met and the user says he doesn't
want to add a key, then tell him that the threshold requirement is not
met and ask him again for a new key.
Don't give him the opportunity to abort the procedure as this will
throw an exception in this case and doesn't bring any value.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>