Once added don't remove a key from signing_keys
There is a requirement from the specification to have at least
threshold (where threshold comes from the trusted current root) number
of keys signing the new root, which were also part of the current
trusted root.
In order to achieve this Lukas suggested that if a key is loaded once
and added to the signing_keys dict, then we shouldn't remove it from
there even if a user wants to remove that key.
This will give us the guarantee that this requirement is met.

This change leads to simplification in that we can use keyids to the
signing_keys instead of key names and this removed the need for the
"_get_rstuf_key_name()" function.

Also, one use case was removed from test_metadata_update where there
are the above threshold requirement is not met as it's practically
impossible to happen now.
Still, I left the exception handling of a possible
"UnsignedMetadataError" error to make sure if there is a problem with
this requirement would have a meaningful message for us.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>