Address multiple comments
I addressed multiple comments in this commit.

First, "_create_keys_table()" no longer requires "root_info" to be
passed as it's a strange to require that.

Second, I added the "deleted" flag for RSTUFKey objects. This allows
to mark keys as deleted and not show them to the user.

Third, I focused on fixing a problem with how we present the keys
information in "_keys_additions()".
There are 3 things we do now:
1. we will tell the user the new root threshold
2. we are showing all of the new root keys to the user
3. we are printing the number of keys required to be added to complete
the threshold requirement

The problem is that not all keys existing in the new root will be used
for signing. This leads to two questions:
1. how do we calculate if the threshold requirement is met
2. how do we explain to the user that some of the keys we are printing
will be used for signing and others will not
3. how do we explain to the user why he does need to add X number of
keys if already has Y number of keys where Y is higher than the
threshold

The solution that I chose was to utilize the signing_keys dictionary,
not mention anything about the threshold, but instead, focus on a number
of keys needed and when I won't show all of the keys, but only those
used for signing (with a proper message before that).
This way we are showing real information about how many keys are needed
to be added, we show only signing keys without the need for any
explanation.

Finally, to make sure that no deleted keys will be shown to the user
I don't use directly the signing_keys dictionary, but I filter it by
removing all keys that were previously marked with the flag "deleted".

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>