Fix two bugs regarding the threshold requirement
There are two bugs I try to address in this pr.

The first bug can be described with a simple example:
Martin can sign root v1, Lukas can sign root v2, and the old and new
threshold is 1 --> As a result root v2 must be signed by both Martin
and Lukas.
The reason is that the old key from the trusted root will be used to
sign the new root version, but it cannot be used to fulfill the
threshold requirement as it's not DEFINED in root v2.

The second bug is that I didn't use keys from trusted root that were
removed for signing. This lead to a situation where the requirement
for a threshold number of keys from the old trusted root signing the
new root is not fulfilled.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>