Remove CLI support for built-in authn/authz
Removes Command Line Interface for the RSTUF API built-in authentication
and authorization (authn/z).

The authn/z will be removed from the RSTUF API, it is related to
https://github.com/repository-service-tuf/repository-service-tuf/issues/491

This commit removes the implementation for `--auth` and `--token` in the
code base and its tests.

Adds consistence to the `--api-url` parameter.

Closes #406

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>