Remove test_generate with invalid type
This test doesn't test anything useful for us as the user will not
have the opportunity to enter an invalid key type as he has to make
a choice from an existing list of valid key types.
If the user writes an invalid key type we know that click will stop
him.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>