Fix multithreading bug in tests
We optimized our unit tests by running them with multiple workers.
This caused an unintended effect because by default metadata
update saves the payload with a default name of
"metadata_update.payload" then caused all of the workers to rewrite
the same file.
That wouldn't be a problem if we didn't read from that file and check
the information inside, but in two of our tests we do just that.

The solution I propose is to use a pytest temp_dir fixture creating
a temp directory. Under that directory, I also added a sub-directory
with the full name of the test to escape other possible collisions
which I noticed were possible.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>