Add alternative admin cli
Added commands:

- `rstuf admin2 ceremony`
- `rstuf admin2 update`
- `rstuf admin2 sign`

Previous work, related discussion, and detailed reasons for the re-write
can be found in #477 and #490. These PRs are superseded by this PR.

Currently, the cli does not interact with the RSTUF API, but reads input
data from files passed as cli arguments (update, sign) and optionally
writes output data to file using the `--payload-out` option (ceremony,
update, sign).

Preliminary API integration can be found in #477. I suggest to
re-implement this in a separate PR, as well as any presentation
improvements discussed in #477.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>