fix: Stop tox `precommit` modifying files
This tox step showed a warning (zero exit code) if the
`.pre-commit-config.yaml` file was out-of-date and
also updated it. When a file is updated `pre-commit`
hooks exit with a non-zero code, whereas we only
wanted to print a warning with this hook.
Also, if you tried to commit again, because the pre-commit
config file had been modified, `pre-commit` did not allow
you to continue with your commit (error: "Your pre-commit
configuration is unstaged.")