Update root files to better represent actual usage
We need to update our root files, so they won't use any of the root
keys as an online key.

Additionally, v2 of the root file shows what it will be like when
a key is removed and a new key is added.
Also, do not use online key that's used in rest of the roles
as a root key.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>