Modify keys used for admin2
There are a few important changes that need to be done so that
those keys can be used by the Worker and the functional tests:
1. Online keys shouldn't be encrypted. They should be in a secure
vault.
2. Online key names should start with their sha256 as that's what is
written on the root.json and the Worker uses that to open their files.
3. We need to rename `ecdsa` to `JJ.ecdsa` for example and do this
for all other `JH`, `JC`, etc. keys as to use the same key names used
in our documentation. When we drop the old keys we would be able
to use the full names - `JanisJoplin`, `JimiHendrix`, `JoeCocker`.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>