Admin2 tests refactoring, key additions and payloads changes (#561)
* Use the same command structure as on the API

On the API we have "/api/v1/metadata/update" and "/api/v1/metadata/sign"
that's why it makes sense to follow the same structure here - "update"
and "sign" will be under "metadata".

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Add new keys to allow realistic key rotation

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Fix small bug not showing online key name

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Update root files to better represent actual usage

We need to update our root files, so they won't use any of the root
keys as an online key.

Additionally, v2 of the root file shows what it will be like when
a key is removed and a new key is added.
Also, do not use online key that's used in rest of the roles
as a root key.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Restructure admin2 tests & update them

Restructure admin2 tests to follow the modules we have for admin2
modules.
Additionally, update them to use the changed payloads and root files.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Use click.prompt for password to ease FT tests

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Modify keys used for admin2

There are a few important changes that need to be done so that
those keys can be used by the Worker and the functional tests:
1. Online keys shouldn't be encrypted. They should be in a secure
vault.
2. Online key names should start with their sha256 as that's what is
written on the root.json and the Worker uses that to open their files.
3. We need to rename `ecdsa` to `JJ.ecdsa` for example and do this
for all other `JH`, `JC`, etc. keys as to use the same key names used
in our documentation. When we drop the old keys we would be able
to use the full names - `JanisJoplin`, `JimiHendrix`, `JoeCocker`.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

* Update tests

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>

---------

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>