Tests: get a better response on tests failures
Now, we got an exception that "file out_file_name cannot be found"
as we try to open a file that doesn't exist, but it doesn't exist
because the execution failed for some reason.
Because of that I make sure to check the stderr before opening the
file and actually receive a meaningfull error.

Additionally, because click merges stdout and stderr by default
(see "mix_stderr" argument https://click.palletsprojects.com/en/8.1.x/api/#testing)
we receive a error "ValueError: stderr not separately captured"
seamingly comming from the subprocess module or another library.
To solve this issue I made "mix_stderr" to True.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>