Introduce "--offline" and "--api-server" to admin
Make "--offline" and "--api-server" admin-level commands which will
be reused by all admin subcommands.
This allows for a single place of exception handling for those two
arguments and consistent behavior across commands.
This allowed the removal of "--bootstrap" in ceremony and "--api-server"
options for both "ceremony" and "sign".

The "--offline" option allows you to have a local setting.SERVER
configuration but still run the commands in a disconnected manner for
testing.

Additionally, with "--api-server" and the "--save" (option for "send")
or "output" (argument for ceremony) you can upload the payload to the
RSTUF API and save the result of the command locally.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>