Change _load_key_from_file_prompt return value
In preparation for supporting multiple online key types, this helper
is changed to return the key and a uri to access the corresponding
signer.

Since the helper is still used for online and offline key configuration,
via a common intermediary helper `_load_key_prompt`, the returned uri is
not yet used.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>