Add '--dry-run' option for 'ceremony' and 'sign'
The '--dry-run' option scope is regarding the commands output.
This means that when the option is used it will not influence the
input options.
Instead it will directly interfere with the output. If '--dry-run'
is used, then the CLI will ignore both '--api-server' admin option
and the RSTUF configuration file and it's corresponding SERVER option.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>